[Modify] LeaveButton, GameStartButton 스크립트 삭제 및 타이틀 캔버스 수정
1) LeaveButton ,GameStartButton 스크립트 Multigame Canvas 스크립트로 병합함 -> 굳이 따로 있을 이유가 없음
2) 타이틀 캔버스에 게임 이름과 터치하라는 텍스트 추가
