+++ Team_Immortal Sprouts_Pummel Party/Assets/Scenes/WooSeok/Lobby Scene 2.unity
@@ -133,7 +133,6 @@ GameObject:
   m_Component:
   - component: {fileID: 179896428}
   - component: {fileID: 179896431}
-  - component: {fileID: 179896430}
   - component: {fileID: 179896429}
   m_Layer: 0
   m_Name: Render Player 2
@@ -190,14 +189,6 @@ MonoBehaviour:
   m_RequiresDepthTexture: 0
   m_RequiresColorTexture: 0
   m_Version: 2
---- !u!81 &179896430
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 179896427}
-  m_Enabled: 1
 --- !u!20 &179896431
 Camera:
   m_ObjectHideFlags: 0
@@ -469,6 +460,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 1340131919436664619, guid: a89076242438ef84e8f5798e3ccc87a2, type: 3}
+      propertyPath: m_RaycastTarget
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 3998255936705922915, guid: a89076242438ef84e8f5798e3ccc87a2, type: 3}
       propertyPath: m_Pivot.x
       value: 0
@@ -565,6 +560,14 @@ PrefabInstance:
       propertyPath: MultiPlayerCanvas
       value: 
       objectReference: {fileID: 1591320725}
+    - target: {fileID: 5905294332163232871, guid: a89076242438ef84e8f5798e3ccc87a2, type: 3}
+      propertyPath: m_CullTransparentMesh
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6898773627669732828, guid: a89076242438ef84e8f5798e3ccc87a2, type: 3}
+      propertyPath: m_RaycastTarget
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a89076242438ef84e8f5798e3ccc87a2, type: 3}
 --- !u!1 &662439875
@@ -577,7 +580,6 @@ GameObject:
   m_Component:
   - component: {fileID: 662439876}
   - component: {fileID: 662439879}
-  - component: {fileID: 662439878}
   - component: {fileID: 662439877}
   m_Layer: 0
   m_Name: Render Player 3
@@ -634,14 +636,6 @@ MonoBehaviour:
   m_RequiresDepthTexture: 0
   m_RequiresColorTexture: 0
   m_Version: 2
---- !u!81 &662439878
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 662439875}
-  m_Enabled: 1
 --- !u!20 &662439879
 Camera:
   m_ObjectHideFlags: 0
@@ -695,7 +689,6 @@ GameObject:
   m_Component:
   - component: {fileID: 772900479}
   - component: {fileID: 772900482}
-  - component: {fileID: 772900481}
   - component: {fileID: 772900480}
   m_Layer: 0
   m_Name: Render Player 4
@@ -752,14 +745,6 @@ MonoBehaviour:
   m_RequiresDepthTexture: 0
   m_RequiresColorTexture: 0
   m_Version: 2
---- !u!81 &772900481
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 772900478}
-  m_Enabled: 1
 --- !u!20 &772900482
 Camera:
   m_ObjectHideFlags: 0
@@ -882,7 +867,7 @@ PrefabInstance:
     - target: {fileID: 5527714054793255316, guid: cb984150341bbb34593c09b27e285b45, type: 3}
       propertyPath: _waitingRoomCanvas
       value: 
-      objectReference: {fileID: 2115222626}
+      objectReference: {fileID: 2115222627}
     - target: {fileID: 5527714054793255317, guid: cb984150341bbb34593c09b27e285b45, type: 3}
       propertyPath: m_RootOrder
       value: 4
@@ -957,7 +942,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1324178387}
   - component: {fileID: 1324178390}
-  - component: {fileID: 1324178389}
   - component: {fileID: 1324178388}
   m_Layer: 0
   m_Name: Render Player 1
@@ -1014,14 +998,6 @@ MonoBehaviour:
   m_RequiresDepthTexture: 0
   m_RequiresColorTexture: 0
   m_Version: 2
---- !u!81 &1324178389
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1324178386}
-  m_Enabled: 1
 --- !u!20 &1324178390
 Camera:
   m_ObjectHideFlags: 0
@@ -1271,11 +1247,6 @@ Transform:
 --- !u!1 &2115222623 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 2487228347764952863, guid: afa6786cc3197624c8891889a2f98d32, type: 3}
-  m_PrefabInstance: {fileID: 2487228348873077568}
-  m_PrefabAsset: {fileID: 0}
---- !u!223 &2115222626 stripped
-Canvas:
-  m_CorrespondingSourceObject: {fileID: 2487228347764952866, guid: afa6786cc3197624c8891889a2f98d32, type: 3}
   m_PrefabInstance: {fileID: 2487228348873077568}
   m_PrefabAsset: {fileID: 0}
 --- !u!114 &2115222627
