--- Team_Immortal Sprouts_Pummel Party/Assets/Scripts/1_Lobby Scene/Lobby/Lobby Canvases/Multigame Canvas/GameStartButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0a42b7bf63e83ff4c843078bbaa88877
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
