--- Team_Immortal Sprouts_Pummel Party/Assets/Scripts/1_Lobby Scene/Lobby/Lobby Canvases/Multigame Canvas/LeaveGame.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class LeaveGame : MonoBehaviour
-{
-    public void OnClickLeaveGameButton()
-    {
-        Application.Quit();
-        Debug.Log("게임을 떠났습니다.");
-    }
-}
