--- Team_Immortal Sprouts_Pummel Party/Assets/Scripts/1_Lobby Scene/Lobby/Lobby Canvases/Multigame Canvas/LeaveGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3f7ab57874f316946a018359a2441112
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
