+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/1_Lobby Scene/Lobby/Title Canvas/Title.cs
@@ -9,32 +9,6 @@ using UnityEngine.InputSystem.EnhancedTouch;
 public class Title : MonoBehaviour
 {
     [SerializeField] private GameObject MultiPlayerCanvas;
-    //private PlayerInput playerInput;
-    //private InputAction touchAction;
-
-    //private void Awake()
-    //{
-    //    playerInput = GetComponent<PlayerInput>();
-    //    touchAction = playerInput.actions["Touch"];
-    //}
-
-    //private void OnEnable()
-    //{
-    //    touchAction.performed -= TouchTitle;
-    //    touchAction.performed += TouchTitle;
-    //}
-
-    //private void OnDisable()
-    //{
-    //    touchAction.performed -= TouchTitle;
-    //}
-
-    //private void TouchTitle(InputAction.CallbackContext value)
-    //{
-    //    gameObject.transform.parent.gameObject.SetActive(false);
-    //    MultiPlayerCanvas.SetActive(true);
-    //    Debug.Log("Touch Title");
-    //}
 
     public void OnTouchTitle()
     {
